pyproject.toml
[project]
dependencies = [
  "click",
  "ibis-framework[duckdb]",
  "fastapi",
  "uvicorn",
  "python-jose",
  "structlog",
  "prometheus-fastapi-instrumentator",
  "jinja2",
  "httpx",
  "numpy",
  "scipy>=1.16.2",
  "pymc>=5.25.1",
]
description = "A vibes CLI for pymc-vibes"
name = "pymc-vibes"
requires-python = ">=3.13"
version = "0.1.0"

[project.scripts]
vibes = "pymc_vibes.cli.main:main"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=68.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src", "tests"]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[tool.pytest.ini_options]
addopts = "-ra -q"
python_files = ["test_*.py", "*_test.py", "*_tests.py"]
testpaths = ["tests"]
